+++ src/components/TarButton.vue
@@ -3,12 +3,12 @@ import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
 import { computed } from "vue";
 
 import TarSpinner from "./TarSpinner.vue";
-import type { ButtonOptions } from "@/types/TarButton";
-import { parseBoolean } from "@/helpers/parsingUtils";
+import type { ButtonOptions } from "../types/TarButton";
+import { parseBoolean } from "../helpers/parsingUtils";
 
 const props = withDefaults(defineProps<ButtonOptions>(), {
   size: "medium",
-  status: "Loading...",
+  status: "Loading…",
   type: "button",
   variant: "primary",
 });