+++ src/components/TarSelect.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import { computed, ref } from "vue";
 
-import type { SelectOptions } from "@/types/TarSelect";
-import { parseBoolean } from "@/helpers/parsingUtils";
+import type { SelectOptions } from "../types/TarSelect";
+import { parseBoolean } from "../helpers/parsingUtils";
 
 const props = withDefaults(defineProps<SelectOptions>(), {
   options: () => [],