+++ src/components/TarSpinner.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
 import { computed } from "vue";
 
-import type { SpinnerOptions } from "@/types/TarSpinner";
-import { parseBoolean } from "@/helpers/parsingUtils";
+import type { SpinnerOptions } from "../types/TarSpinner";
+import { parseBoolean } from "../helpers/parsingUtils";
 
 const props = withDefaults(defineProps<SpinnerOptions>(), {
-  label: "Loading...",
+  label: "Loading…",
   role: "status",
 });