+++ .storybook/main.ts
@@ -3,7 +3,7 @@ import type { StorybookConfig } from "@storybook/vue3-vite";
 const config: StorybookConfig = {
   stories: ["../src/**/*.mdx", "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
   staticDirs: ["../public"],
-  addons: ["@storybook/addon-links", "@storybook/addon-essentials", "@storybook/addon-interactions"],
+  addons: ["@storybook/addon-links", "@storybook/addon-essentials", "@storybook/addon-interactions", "@chromatic-com/storybook"],
   framework: {
     name: "@storybook/vue3-vite",
     options: {},