+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+Nothing yet.
+
+## [2.1.0] - 2024-04-17
+
 ### Added
 
 - Build workflow `pull_request` trigger on `main` branch.
@@ -86,7 +90,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implemented an Image component.
 - Implemented an Input component.
 
-[unreleased]: https://github.com/Logitar/Vue3Ui/compare/v2.0.5...HEAD
+[unreleased]: https://github.com/Logitar/Vue3Ui/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/Logitar/Vue3Ui/compare/v2.0.5...v2.1.0
 [2.0.5]: https://github.com/Logitar/Vue3Ui/compare/v2.0.4...v2.0.5
 [2.0.4]: https://github.com/Logitar/Vue3Ui/compare/v2.0.3...v2.0.4
 [2.0.3]: https://github.com/Logitar/Vue3Ui/compare/v2.0.2...v2.0.3